Refactored lammps parsing.
