+++ src/diffusion_for_multi_scale_molecular_dynamics/active_learning_loop/lammps/outputs.py
